pyproject.toml
[tool.poetry]
name = "officialWebsite"
version = "0.1.0"
description = "Poetry file for officialWebsite DSC TIET"
authors = ["DSC TIET"]

[tool.poetry.dependencies]
python = "^3.7, <3.9"
Django = "2.2.10"
django-cleanup = "4.0.0"
django-cors-headers = "3.3.0"
django-environ = "0.4.5"
django-filter = "2.2.0"
djangorestframework = "3.10.3"
Markdown = "3.2.2"
Pillow = "7.1.1"
pre-commit = "2.4.0"
requests = "^2.23.0"
protobuf = "3.11.3"
django-storages = {extras = ["google"], version = "^1.9.1"}
google-cloud-storage = "^1.28.1"

[tool.black]
line-length = 200
target-version = ['py36', 'py37', 'py38']
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.*
  | \.git
  | \.mypy_cache
  | \.venv
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''